Refactor - Simplify OpenStaticAnalyzer
